+++ tests/kotsu/components/Breadcrumb.test.js
@@ -1,7 +1,7 @@
 /* eslint-env jest */
 
 import cloneDeep from 'lodash/cloneDeep'
-import { renderString } from '../../utils/nunjucks'
+import { renderString } from '../../../modules/nunjucks-test-utils'
 
 const render = (template, context = mockContext, parse) => renderString(template, context, parse)
 const mockContext = {