+++ tests/utils/scriptsData.js
@@ -1,4 +1,0 @@
-const scriptsData = require(`../../source/data/scripts`)
-const { grunt } = require(`../utils/grunt`)
-
-module.exports = scriptsData(grunt)